separate addItem logic into separate hook
